Move WPLineItem and associated components to own file
